half version

+++ unity/Assets/Shader/libDecoder.cginc
@@ -14,21 +14,21 @@ contact: peng_huailiang@qq.com
 #include "libActive.cginc"
 
 #define DefineDecodeBuffer(seq)	\
-	RWStructuredBuffer<float> decoder_conv##seq##;	\
-	RWStructuredBuffer<float> decoder_conv##seq##_conved;		\
-	RWStructuredBuffer<float> decoder_conv##seq##_statistic;	\
+	RWStructuredBuffer<half> decoder_conv##seq##;	\
+	RWStructuredBuffer<half> decoder_conv##seq##_conved;		\
+	RWStructuredBuffer<half> decoder_conv##seq##_statistic;	\
 
 
 #define DefineResiduleConv(id, input, output, seq, idx) \
 	for(int j = 0;j < depth; j++) \
 	{ 	\
-		float v = 0.0f;	\
+		half v = 0.0f;	\
 		for(int i= 0; i < depth; i++)	\
 		{	\
-			float3x3 xsamp = StdSample(decoder_conv0, id.x, id.y, i, input, depth);	\
-			float3x3 kernel = decoder_g_r##seq##_c##idx##_Conv_weights[depth * i + j];	\
-			float3x3 conv = xsamp * kernel;	\
-			float3 iall = conv[0] + conv[1] + conv[2];	\
+			half3x3 xsamp = StdSample(decoder_conv0, id.x, id.y, i, input, depth);	\
+			half3x3 kernel = decoder_g_r##seq##_c##idx##_Conv_weights[depth * i + j];	\
+			half3x3 conv = xsamp * kernel;	\
+			half3 iall = conv[0] + conv[1] + conv[2];	\
 			v += iall[0] + iall[1] + iall[2];	\
 		}	\
 		int indx = (output * depth) * id.x + depth * id.y + j;	\
@@ -38,26 +38,26 @@ contact: peng_huailiang@qq.com
 
 #define DeifineResiduleInst(id, seq, r)	\
 	int indx = StdID(id, width, depth);	\
-	float color = input_writable[indx];	\
-	float mean = decoder_conv0_statistic[id.z * 2];	\
-	float variance = decoder_conv0_statistic[id.z * 2 + 1];	\
-	float inv = rsqrt(variance + EPSILON);	\
-	float normalized = (color - mean) * inv;	\
-	float scale = decoder_g_r##seq##_bn##r##_scale[id.z];	\
-	float offset = decoder_g_r##seq##_bn##r##_offset[id.z];	\
+	half color = input_writable[indx];	\
+	half mean = decoder_conv0_statistic[id.z * 2];	\
+	half variance = decoder_conv0_statistic[id.z * 2 + 1];	\
+	half inv = rsqrt(variance + EPSILON);	\
+	half normalized = (color - mean) * inv;	\
+	half scale = decoder_g_r##seq##_bn##r##_scale[id.z];	\
+	half offset = decoder_g_r##seq##_bn##r##_offset[id.z];	\
 	input_writable[indx] = scale * normalized + offset;		
 
 
 #define DefineDecoderConv(id, width, depth1, depth2, idx)	\
 	for(uint j = 0; j < depth2; j++) \
 	{ 	\
-		float v = 0.0f;	\
+		half v = 0.0f;	\
 		for(uint i = 0; i < depth1; i++)	\
 		{	\
-			float3x3 xsamp = StdSlowSample(decoder_conv##idx##_conved, id.x, id.y, i, width, depth1);	\
-			float3x3 kernel = decoder_g_d##idx##_dc_conv2d_Conv_weights[depth2 * i + j];	\
-			float3x3 conv = xsamp * kernel;	\
-			float3 iall = conv[0] + conv[1] + conv[2];	\
+			half3x3 xsamp = StdSlowSample(decoder_conv##idx##_conved, id.x, id.y, i, width, depth1);	\
+			half3x3 kernel = decoder_g_d##idx##_dc_conv2d_Conv_weights[depth2 * i + j];	\
+			half3x3 conv = xsamp * kernel;	\
+			half3 iall = conv[0] + conv[1] + conv[2];	\
 			v += iall[0] + iall[1] + iall[2];	\
 		}	\
 		int indx = (width * depth2) * id.x + depth2 * id.y + j;	\
@@ -67,13 +67,13 @@ contact: peng_huailiang@qq.com
 
 #define DefineDecoderInstRelu(id, seq)	\
 	int inx = StdID(id, width, depth);	\
-	float color = decoder_conv##seq##[inx];	\
-	float mean = decoder_conv##seq##_statistic[id.z * 2];	\
-	float variance = decoder_conv##seq##_statistic[id.z * 2 + 1];	\
-	float inv = rsqrt(variance + EPSILON);	\
-	float normalized = (color - mean) * inv;	\
-	float scale = decoder_g_d##seq##_bn_scale[id.z];	\
-	float offset = decoder_g_d##seq##_bn_offset[id.z];	\
+	half color = decoder_conv##seq##[inx];	\
+	half mean = decoder_conv##seq##_statistic[id.z * 2];	\
+	half variance = decoder_conv##seq##_statistic[id.z * 2 + 1];	\
+	half inv = rsqrt(variance + EPSILON);	\
+	half normalized = (color - mean) * inv;	\
+	half scale = decoder_g_d##seq##_bn_scale[id.z];	\
+	half offset = decoder_g_d##seq##_bn_offset[id.z];	\
 	decoder_conv##seq##[inx] = relu(scale * normalized + offset);
 
 
@@ -90,7 +90,7 @@ contact: peng_huailiang@qq.com
 
 #define DefineDecoderExpand(id, idx, pidx) \
 	int indx = StdID(id, width, depth);	\
-	float v = decoder_conv##pidx##[indx];	\
+	half v = decoder_conv##pidx##[indx];	\
 	DefineDecoderPad(id, idx)
